--- front-end/src/Item.js
@@ -1,23 +1,0 @@
-import React from "react"
-import { dot, chipDays, chipAmount } from "./itemColoring";
-import "./MyFridge.css";
-
-const Item = (props) => {
-    return (
-        // elements for the table
-        <tbody>
-            <tr>
-                <td>
-                    {/* we will use the functions from itemColoring */}
-                    {/* and the css from MyFridge.css */}
-                    <span>{dot(props.food.daysLeft)}</span>
-                    <span className="title">{props.food.title}</span>
-                    <span>{chipAmount(props.food.amount)}</span>
-                    <span>{chipDays(props.food.daysLeft)}</span>
-                </td>
-            </tr>
-        </tbody>
-    )
-}
-
-export default Item
