+++ front-end/src/Recommendations/Recommendations.js
@@ -1,4 +1,4 @@
-import NavBar from "./NavBar";
+import NavBar from "../NavBar";
 import Button from "@material-ui/core/Button";
 import Grid from "@material-ui/core/Grid";
 import "./Recommendations.css";
@@ -6,9 +6,10 @@ import ButtonBase from '@material-ui/core/ButtonBase';
 import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import { Link } from "react-router-dom";
-import { recipes } from "./recommendationDataFake";
+import { useState } from "react";
+import { recipes } from "../data/recommendationDataFake";
+import {ReadyToMake} from "./ReadyToMake";
 import {SavedRecipes} from "./SavedRecipes";
-import {Recommendations} from "./Recommendations";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -28,11 +29,13 @@ const useStyles = makeStyles((theme) => ({
     },
   }));
 
-const ReadyToMake = (props) => {
+const Recommendations = (props) => {
   const classes = useStyles();
 
+  
 
   const RecItem = (data) => {
+    const [buttonText,setButtonText] = useState('Save Recipe')
     return(
       <div className={classes.root}>
           <Grid container spacing={0} >
@@ -44,7 +47,7 @@ const ReadyToMake = (props) => {
               <Grid item xs = {6}>
                   <div className = "recommendations-recipeTitle">{data.title}</div>
                   <div className = "recommendations-info">{data.ingredients} Ingredients | {data.minutes} Minutes</div>
-                  <Button className = "recommendations-addButton" variant = "outlined" color="primary"> Save Recipe </Button>
+                  <Button className = "recommendations-addButton" variant = "outlined" color="primary" onClick={() => setButtonText('Saved!')}> {buttonText} </Button>
               </Grid> 
           </Grid>
     </div>
@@ -57,12 +60,12 @@ const ReadyToMake = (props) => {
     <header className="App-header">
       
     <Link to = "/Recommendations" className = "recommendations-link">
-      <Button className = "recommendations-unusedButton">
+      <Button className = "recommendations-usedButton">
         Recipes of the Day
       </Button>
     </Link>
       <Link to = "/ReadyToMake" className = "recommendations-link">
-        <Button className = "recommendations-usedButton">
+        <Button className = "recommendations-unusedButton">
           Ready to Make
         </Button>
       </Link>
@@ -84,4 +87,4 @@ const ReadyToMake = (props) => {
 );
   };
 
-export { ReadyToMake };
+export { Recommendations };
