+++ front-end/src/StorageTimeSearch/Item.css
@@ -26,7 +26,6 @@
   margin-left: auto;
   margin-right: auto;
   padding: 10px;
-  position: sticky;
   width: 85%;
   background-color: white;
   border-bottom: solid;
